peel off manifest into its own file
